+++ client/src/views/Awx.vue
@@ -16,8 +16,10 @@
           </nav>         
           <div class="columns">
             <div class="column">
+              <BulmaCheckbox checktype="checkbox" v-model="awx.use_credentials" label="Use Credentials" />
               <BulmaInput icon="globe" v-model="awx.uri" label="Uri" placeholder="https://awx.domain.local" :required="true" :hasError="$v.awx.uri.$invalid" :errors="[]" />
-              <BulmaInput icon="lock" type="password" v-model="awx.token" label="Token" placeholder="Token" :required="true" :hasError="$v.awx.token.$invalid" :errors="[]" />
+              <BulmaInput icon="user" v-if="awx.use_credentials" v-model="awx.username" label="Username" placeholder="Username" :required="true" :hasError="$v.awx.username.$invalid" :errors="[]" />
+              <BulmaInput icon="lock" type="password" v-model="awx.token" :label="(awx.use_credentials)?'Password':'Token'" :required="true" :hasError="$v.awx.token.$invalid" :errors="[]" />
             
             </div>
             <div class="column">
@@ -56,7 +58,9 @@
           awx:{
             uri:"",
             token:"",
+            username:"",
             ignore_certs:true,
+            use_credentials:false,
             ca_bundle:""
           }
         }
@@ -114,7 +118,12 @@
           required:requiredIf(function(awx){
             return !awx.ignore_certs
           })
-        }
+        },
+        username:{
+          required:requiredIf(function(awx){
+            return awx.use_credentials
+          })
+        }        
       }
     },
     mounted() { // when the Vue app is booted up, this is run automatically.